refactor : 파일 분리, 에러 데이터 생성
