+++ src/recipe/type/error.ts
@@ -1,0 +1,5 @@
+const RecipeError = {
+	RECIPE_NOT_FOUND: 'RECIPE_NOT_FOUND',
+};
+
+export default RecipeError;
